+++ OrderService.Domain/DomainModel/OrderDomainModel/Commands/PlaceOrderCommand.cs
@@ -17,7 +17,7 @@ namespace OrderService.Domain.DomainModel.OrderDomainModel.Commands
         public PlaceOrderCommand(
             OrderId id,
             string userId,
-            int orderTotal,
+            decimal orderTotal,
             DateTime orderPlaced)
             : base(id)
         {
@@ -32,7 +32,7 @@ namespace OrderService.Domain.DomainModel.OrderDomainModel.Commands
 
         public string UserId { get; }
 
-        public int OrderTotal { get; }
+        public decimal OrderTotal { get; }
 
         public DateTime OrderPlaced { get; }
